Correct signature of usleep() method
git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk@710216 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/cpp/src/qpid/sys/windows/Time.cpp
@@ -80,7 +80,7 @@ void sleep(int secs) {
     ::Sleep(secs * 1000);
 }
 
-void usleep(uint32_t usecs) {
+void usleep(uint64_t usecs) {
     DWORD msecs = usecs / 1000;
     if (msecs == 0)
         msecs = 1;
